tracklab-main/pyproject.toml
[tool.poetry]
name = "tracklab"
version = "1.1.2"
description = ""
authors = ["Vladimir Somers <vladimir.somers@uclouvain.be>", "Victor Joos <victor.joos@uclouvain.be>", "Baptiste Standaert <baptiste.standaert@uclouvain.be>"]
readme = "README.md"
packages = [{include = "tracklab"}]

[tool.poetry.dependencies]
python = "^3.9"
# openpifpaf = {git="https://github.com/PbTrack/openpifpaf", branch="pbtrack", develop=true, extras=["dev","train","test"]}
poseval = {git="https://github.com/TrackingLaboratory/poseval", branch="pbtrack", develop=true}
posetrack21 = {path = "plugins/eval/PoseTrack21/posetrack21", develop = true}
posetrack21_mot = {path = "plugins/eval/PoseTrack21/posetrack21_mot", develop = true}
hydra-core = "^1.3"
lightning = "^2.0"
pytorch_lightning = "^2.0"
numpy = "^1.23.5"
openmim = "^0.3.9"
ultralytics = "8.0.61"
sphinx = "^7.2"
sphinx_rtd_theme = "^2.0"
myst-parser = "^2.0"
filterpy = "^1.4.5"
torch = "1.13.1"
torchvision = "0.14.1"
mmdet = "~3.1.0"
chumpy = "0.66"
lap = {git = "https://github.com/TrackingLaboratory/lap"}
# track-bench-track = {path = "plugins/track", develop = true}
track-bench-track = "1.0.1"
mmengine = "^0.10.1"
timm = "^0.9.12"
torchreid = {git = "https://github.com/VlSomers/bpbreid"}
openpifpaf = "^0.13.11"
soccernet = "^0.1.53"
mmpose = "^1.2.0"
trackeval = {git = "https://github.com/SoccerNet/sn-trackeval.git"}
yt-dlp = ">2023.12.30"
gdown = "^4.7.1"
pandas = "^2.1.0"

[tool.poetry.scripts]
tracklab = 'tracklab.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"